requirements.txt
async-generator==1.10
attrs==23.1.0
bidict==0.22.1
blinker==1.6.2
cairocffi==1.6.1
CairoSVG==2.7.1
certifi==2023.5.7
cffi==1.16.0
charset-normalizer==3.1.0
click==8.1.3
cssselect2==0.7.0
defusedxml==0.7.1
driver==0.1.0
exceptiongroup==1.1.1
Flask==2.3.2
Flask-Cors==3.0.10
Flask-SocketIO==5.3.6
fonttools==4.42.1
fpdf2==2.7.5
h11==0.14.0
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
looseversion==1.2.0
MarkupSafe==2.1.3
netifaces==0.10.6
networkx==3.1
nibabel==5.1.0
numpy==1.24.3
opencv-python==4.8.0.76
outcome==1.2.0
packaging==23.1
pathlib==1.0.1
Pillow==10.0.1
pycparser==2.21
pydot==1.4.2
PyMuPDF==1.23.3
PyMuPDFb==1.23.3
pyparsing==3.0.9
PyPDF2==3.0.1
pyperclip==1.8.2
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.7.1
python-socketio==5.9.0
pytz==2023.3
reportlab==4.0.5
requests==2.31.0
scipy==1.11.2
selenium==4.12.0
simple-websocket==0.10.1
simplejson==3.19.1
six==1.16.0
sniffio==1.3.0
socketio==0.2.1
sortedcontainers==2.4.0
tinycss2==1.2.1
tqdm==4.65.0
traits==6.3.2
trio==0.22.0
trio-websocket==0.10.3
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.6
webdriver-manager==3.8.6
webencodings==0.5.1
websockets==11.0.3
Werkzeug==2.3.6
wsproto==1.2.0
zipp==3.15.0
